HBX-1589: Move class 'org.hibernate.cfg.JDBCBinder' to 'org.hibernate.tool.internal.reveng.JdbcBinder'
